requirements.txt
gradio>=4.0
requests
pygltflib
python-dotenv
psutil
